+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Enhancements
 
+* Added support for coercion with `add`, `<<` and `merge` when mutating Set.
 * Added support for coercion with `insert` when mutating Array.
 * Added support for coercion with `merge!` and `store` when mutating Hash.